+++ clients/email-template-client/src/openapi.d.ts
@@ -969,6 +969,7 @@ declare namespace Paths {
     }
 }
 
+
 export interface OperationMethods {
   /**
    * saveTemplate - saveTemplate
@@ -1134,6 +1135,7 @@ export interface PathsDictionary {
 }
 
 export type Client = OpenAPIClient<OperationMethods, PathsDictionary>
+
 
 export type ApproveAction = Components.Schemas.ApproveAction;
 export type ApprovedAt = Components.Schemas.ApprovedAt;