+++ clients/message-client/src/openapi.d.ts
@@ -1173,6 +1173,12 @@ declare namespace Paths {
                 /**
                  * Total of unread messages
                  * example:
+                 * 14
+                 */
+                count: number;
+                /**
+                 * Total of unread messages
+                 * example:
                  * 0
                  */
                 unread?: number;
@@ -1371,6 +1377,151 @@ declare namespace Paths {
                  * Matched threads ids
                  */
                 results: string[];
+            }
+            export interface $403 {
+            }
+        }
+    }
+    namespace SearchMessages {
+        export type RequestBody = Components.Schemas.SearchParamsV2;
+        namespace Responses {
+            export interface $200 {
+                /**
+                 * Total of matched messages
+                 * example:
+                 * 14
+                 */
+                hits: number;
+                /**
+                 * Matched messages
+                 */
+                results: {
+                    /**
+                     * Entity ID
+                     * example:
+                     * 3fa85f64-5717-4562-b3fc-2c963f66afa6
+                     */
+                    _id: string;
+                    /**
+                     * Entity title
+                     */
+                    _title: string;
+                    /**
+                     * Organization ID the entity belongs to
+                     * example:
+                     * 206801
+                     */
+                    _org: string;
+                    /**
+                     * URL-friendly identifier for the entity schema
+                     * example:
+                     * message
+                     */
+                    _schema: string;
+                    /**
+                     * Entity tags
+                     * example:
+                     * [
+                     *   "pricing",
+                     *   "INBOX"
+                     * ]
+                     */
+                    _tags?: string[];
+                    /**
+                     * Created date
+                     * example:
+                     * 2021-02-09T12:41:43.662Z
+                     */
+                    _created_at: string; // date-time
+                    /**
+                     * Updated date
+                     * example:
+                     * 2021-02-10T09:14:31.990Z
+                     */
+                    _updated_at: string; // date-time
+                    /**
+                     * Message ID which is from email provider. If you provide `message-id`, API overrides by its own value.
+                     * example:
+                     * <0102017b97a502f8-a67f01c2-68cc-4928-b91b-45853f34e259-000000@eu-west-1.amazonses.com>
+                     */
+                    message_id?: string;
+                    /**
+                     * User ID of user sends the message.
+                     * example:
+                     * 206801
+                     */
+                    sender?: string;
+                    /**
+                     * Subject
+                     * example:
+                     * Request for solar panel price
+                     */
+                    subject: string;
+                    /**
+                     * HTML body
+                     * example:
+                     * <div>We at ABC GmbH would like to request a price quote for the solar panel.</div>
+                     */
+                    html?: string;
+                    /**
+                     * Text body
+                     * example:
+                     * We at ABC GmbH would like to request a price quote for the solar panel.
+                     */
+                    text?: string;
+                    from: Components.Schemas.Address;
+                    reply_to?: Components.Schemas.Address;
+                    /**
+                     * To email addresses
+                     */
+                    to?: Components.Schemas.Address[];
+                    /**
+                     * Cc email addresses
+                     */
+                    cc?: Components.Schemas.Address[];
+                    /**
+                     * Bcc email addresses
+                     */
+                    bcc?: Components.Schemas.Address[];
+                    file?: /* Message attachments */ Components.Schemas.AttachmentsRelation;
+                    /**
+                     * References header. Value is the series of `message_id` which is reparated by space to indicate that message has parent.            The last message ID in references identifies the parent. The first message ID in references identifies the first message in the thread.            The basic idea is that sender should copy `references` from the parent and append the parent's `message_id` when replying.
+                     *
+                     * example:
+                     * <0102017b97a502f8-a67f01c2-68cc-4928-b91b-45853f34e259-000000@eu-west-1.amazonses.com> <CALHgQpziyxW9NaFUs+nRMykzr6Ljq6vjq4WO9SaihAuMasuDyg@mail.gmail.com>
+                     */
+                    references?: string;
+                    /**
+                     * In-Reply-To header. Value is the `message_id` of parent message.
+                     *
+                     * example:
+                     * <CALHgQpziyxW9NaFUs+nRMykzr6Ljq6vjq4WO9SaihAuMasuDyg@mail.gmail.com>
+                     */
+                    in_reply_to?: string;
+                    /**
+                     * User ID of user read the message.
+                     */
+                    user_read_message?: string[];
+                    /**
+                     * Organization ID of organization read the message.
+                     */
+                    org_read_message?: string[];
+                    /**
+                     * Sent message status. The array contains sending message status corresponding to all recipients. For more detail, check `send_status` of each recipient in `to`, `cc`, `bcc`            Reference at <https://docs.aws.amazon.com/ses/latest/DeveloperGuide/monitor-sending-activity.html>
+                     *
+                     */
+                    send_status?: ("SEND" | "DELIVERY" | "REJECT" | "COMPLAINT" | "BOUNCE" | "ERROR")[];
+                    /**
+                     * Message type
+                     */
+                    type?: "SENT" | "RECEIVED";
+                    /**
+                     * Template ID used for sending message.
+                     * example:
+                     * 3f34ce73-089c-4d45-a5ee-c161234e41c3
+                     */
+                    template_id?: string;
+                }[];
             }
             export interface $403 {
             }
@@ -2110,6 +2261,7 @@ declare namespace Paths {
     }
 }
 
+
 export interface OperationMethods {
   /**
    * updateMessage - updateMessage
@@ -2151,6 +2303,16 @@ export interface OperationMethods {
     data?: any,
     config?: AxiosRequestConfig  
   ): OperationResponse<Paths.DeleteMessage.Responses.$204>
+  /**
+   * searchMessages - searchMessages
+   * 
+   * Search Messages
+   */
+  'searchMessages'(
+    parameters?: Parameters<UnknownParamsObject> | null,
+    data?: Paths.SearchMessages.RequestBody,
+    config?: AxiosRequestConfig  
+  ): OperationResponse<Paths.SearchMessages.Responses.$200>
   /**
    * trashMessage - trashMessage
    * 
@@ -2602,6 +2764,18 @@ export interface PathsDictionary {
       data?: any,
       config?: AxiosRequestConfig  
     ): OperationResponse<Paths.DeleteMessage.Responses.$204>
+  }
+  ['/v1/message/messages:search']: {
+    /**
+     * searchMessages - searchMessages
+     * 
+     * Search Messages
+     */
+    'post'(
+      parameters?: Parameters<UnknownParamsObject> | null,
+      data?: Paths.SearchMessages.RequestBody,
+      config?: AxiosRequestConfig  
+    ): OperationResponse<Paths.SearchMessages.Responses.$200>
   }
   ['/v1/message/messages/{id}/trash']: {
     /**
@@ -3085,6 +3259,7 @@ export interface PathsDictionary {
 }
 
 export type Client = OpenAPIClient<OperationMethods, PathsDictionary>
+
 
 export type Address = Components.Schemas.Address;
 export type AttachmentsRelation = Components.Schemas.AttachmentsRelation;