+++ clients/metering-client/src/openapi.d.ts
@@ -1185,6 +1185,7 @@ declare namespace Paths {
     }
 }
 
+
 export interface OperationMethods {
   /**
    * getCustomerMeters - Get Customer Meters
@@ -1522,6 +1523,7 @@ export interface PathsDictionary {
 }
 
 export type Client = OpenAPIClient<OperationMethods, PathsDictionary>
+
 
 export type ActionLabel = Components.Schemas.ActionLabel;
 export type ActivityId = Components.Schemas.ActivityId;