+++ clients/erp-integration-client/CHANGELOG.md
@@ -1,1 +1,7 @@
 # Changelog
+
+## 0.4.0
+
+### Minor Changes
+
+- Minor version bump for all client packages